+++ packages/core/src/icons/components/Add.js
@@ -1,0 +1,15 @@
+import * as React from 'react';
+
+function SvgAdd(props) {
+    return (
+        <svg width="1em" height="1em" viewBox="0 0 32 32" {...props}>
+            <path
+                data-name="\u89D2\u4E38\u9577\u65B9\u5F62 1 \u306E\u30B3\u30D4\u30FC 8"
+                d="M22 17h-5v5a1 1 0 11-2 0v-5h-5a1 1 0 010-2h5v-5a1 1 0 012 0v5h5a1 1 0 010 2z"
+                fillRule="evenodd"
+            />
+        </svg>
+    );
+}
+
+export default SvgAdd;