+++ packages/core/src/icons/components/CrmGender.js
@@ -1,0 +1,15 @@
+import * as React from 'react';
+
+function SvgCrmGender(props) {
+    return (
+        <svg width="1em" height="1em" viewBox="0 0 1000 1000" {...props}>
+            <path
+                fillRule="evenodd"
+                clipRule="evenodd"
+                d="M500.1 187.5h-.1c-155.3 0-281.2 125.9-281.2 281.3C218.8 624.1 344.7 750 500 750h.1V187.5zm-.1 625c-189.8 0-343.7-153.9-343.7-343.7C156.3 278.9 310.2 125 500 125c189.8 0 343.8 153.9 343.8 343.8 0 189.8-154 343.7-343.8 343.7z"
+            />
+        </svg>
+    );
+}
+
+export default SvgCrmGender;