+++ packages/core/src/icons/components/Dropdown.js
@@ -1,0 +1,16 @@
+import * as React from 'react';
+
+function SvgDropdown(props) {
+    return (
+        <svg viewBox="-2148 -4721 32 32" width="1em" height="1em" {...props}>
+            <path
+                data-name="\u9577\u65B9\u5F62 7893"
+                fill="transparent"
+                d="M-2148-4721h32v32h-32z"
+            />
+            <path data-name="\u30D1\u30B9 5" d="M-2139-4709h14l-7 10z" />
+        </svg>
+    );
+}
+
+export default SvgDropdown;