+++ packages/core/src/icons/components/Loading.js
@@ -1,0 +1,26 @@
+import * as React from 'react';
+
+function SvgLoading(props) {
+    return (
+        <svg
+            id="loading_svg__loading_svg"
+            width="1em"
+            height="1em"
+            viewBox="0 0 32 32"
+            {...props}
+        >
+            <defs>
+                <style />
+            </defs>
+            <path
+                id="loading_svg__\u6955\u5186\u5F62_1_\u306E\u30B3\u30D4\u30FC"
+                data-name="\u6955\u5186\u5F62 1 \u306E\u30B3\u30D4\u30FC"
+                d="M4329.07 962.108A5.994 5.994 0 104334 968a5.429 5.429 0 00-.11-1.072l1.98-.358a8.4 8.4 0 01.13 1.43 8.037 8.037 0 11-6.57-7.865z"
+                transform="translate(-4312 -952)"
+                fillRule="evenodd"
+            />
+        </svg>
+    );
+}
+
+export default SvgLoading;