+++ packages/core/src/icons/components/Lock.js
@@ -1,0 +1,15 @@
+import * as React from 'react';
+
+function SvgLock(props) {
+    return (
+        <svg width="1em" height="1em" viewBox="0 0 32 32" {...props}>
+            <path
+                data-name="\u89D2\u4E38\u9577\u65B9\u5F62 10 \u306E\u30B3\u30D4\u30FC 6"
+                d="M23.5 24h-15a.5.5 0 01-.5-.5v-9a.5.5 0 01.5-.5H10v-1a6 6 0 0112 0v1h1.5a.5.5 0 01.5.5v9a.5.5 0 01-.5.5zM20 13a4 4 0 10-8 0v2h2v-2a2 2 0 014 0v2h2v-2zm2 3H10v6h12v-6z"
+                fillRule="evenodd"
+            />
+        </svg>
+    );
+}
+
+export default SvgLock;