+++ packages/core/src/icons/components/PrevPage.js
@@ -1,0 +1,19 @@
+import * as React from 'react';
+
+function SvgPrevPage(props) {
+    return (
+        <svg viewBox="-1612 -5337 32 32" width="1em" height="1em" {...props}>
+            <path
+                data-name="\u9577\u65B9\u5F62 9343"
+                fill="transparent"
+                d="M-1612-5337h32v32h-32z"
+            />
+            <path
+                data-name="\u5408\u4F53 515"
+                d="M-1601.799-5321l8.485-8.486 1.415 1.414-7.071 7.071 7.07 7.071-1.414 1.414z"
+            />
+        </svg>
+    );
+}
+
+export default SvgPrevPage;