+++ packages/core/src/icons/components/RemoveElement.js
@@ -1,0 +1,19 @@
+import * as React from 'react';
+
+function SvgRemoveElement(props) {
+    return (
+        <svg viewBox="-2148 -4777 32 32" width="1em" height="1em" {...props}>
+            <path
+                data-name="\u9577\u65B9\u5F62 7838"
+                fill="transparent"
+                d="M-2148-4777h32v32h-32z"
+            />
+            <path
+                data-name="\u5408\u4F53 508"
+                d="M-2143-4761.001a11.015 11.015 0 0111-11 11.014 11.014 0 0111 11 11.011 11.011 0 01-11 11 11.012 11.012 0 01-11-11zm2 0a9.007 9.007 0 009 9 9.009 9.009 0 009-9 9.011 9.011 0 00-9-9 9.009 9.009 0 00-9 9zm3 1v-2h12v2z"
+            />
+        </svg>
+    );
+}
+
+export default SvgRemoveElement;